add `npm start` command
- builds website from source
- starts a local http-server
- opens website in browser
- monitor & re-build modified source files
