+++ package.json
@@ -4,9 +4,12 @@
   "version": "0.0.1",
   "type": "module",
   "scripts": {
-    "preview": "npm run build && npm run serve",
+    "start": "npm run build && run-p autobuild serve",
     "serve": "http-server dist/ -c-1 -p 0 -o",
-    "build": "npm run clean && node scripts/build.mjs",
+    "build": "npm run clean && run-p build-images build-other",
+    "autobuild": "node scripts/autobuild.mjs",
+    "build-images": "node scripts/build-images.mjs",
+    "build-other": "node scripts/build-other.mjs",
     "clean": "node scripts/clean.mjs"
   },
   "repository": {
@@ -18,10 +21,12 @@
     "@svgdotjs/svg.js": "3.1.2"
   },
   "devDependencies": {
+    "chokidar": "3.5.3",
     "http-server": "14.1.0",
-    "tiff-to-png": "2.0.2",
+    "npm-run-all": "4.1.5",
     "shelljs": "0.8.5",
     "svgdom": "0.1.10",
+    "tiff-to-png": "2.0.2",
     "ts-loader": "9.3.0",
     "typescript": "4.6.4",
     "webpack": "5.72.0",
