+++ scripts/build-images.mjs
@@ -1,0 +1,17 @@
+import shell from "shelljs";
+import ConvertTiff from "tiff-to-png";
+import glob from "glob";
+
+const image_dir = "./dist/fileadmin/templates/iwr_vis";
+shell.mkdir("-p", image_dir);
+
+// convert tif images to png & write to dist/fileadmin/iwr_vis folder
+const options = {
+  cwd: ".",
+};
+const converter = new ConvertTiff();
+const forFiles = (err, files) => {
+  if (err) console.log(err);
+  converter.convertArray(files, image_dir).then(console.log);
+};
+glob("**/img/*.tif", options, forFiles);
