+++ scripts/build-other.mjs
@@ -1,0 +1,5 @@
+import { buildStatic } from "./build-static.mjs";
+import { buildScripts } from "./build-scripts.mjs";
+
+buildStatic();
+buildScripts();
